refactor(schemas): move test files into tests/ subfolder
